Configure SXC 9.2 for nuke's XcSitecore target
As 9.2 uses WDP packages it is possible to install the Sitecore Commerce packages "offline".
Created a new Nuke target 'XcSitecore' and use SIF to install the necessary tasks for the 'sitecore' container.
This has been done by extracting the 'includes' needed for the 'sitecore' container from the Master_SingleServer.json SIF configuration